Implement i18n. Closes #28
Not fully happy with the TS implementation of the `translation` function. I wanted
to make sure that the app won't crash if an unknown key is passed. Typescript saves
me from doing that, but then all the possible keys should always be implemented in
all languages, or it would fail.

Now there is a fallback in case a key is not found for the current locale, and it works
as follows:

* If `$default` has been provided in the params, it falls back to that
* Otherwise it falls back to the key in the default locale
* Otherwise it falls back to the key itself

However, TS is not quite aware of that. And so to even use the `$default` param you
have to trick it. See, e.g. `Interval.tsx`